+++ api/Application/Products/Commands/UpdateProduct/UpdateProductDto.cs
@@ -21,10 +21,10 @@ public class UpdateProductDto : IMapFrom<Product>
     
     public string? PictureUrl { get; set; }
 
-    [Range(1, int.MaxValue)]
+    [Range(1.0, double.MaxValue)]
     public decimal Price { get; init; }
 
-    [Range(1, int.MaxValue)]
+    [Range(1.0, double.MaxValue)]
     public int QuantityOnStock { get; set; }
 
     public void Mapping(Profile profile)
